Refactor the way exceptions are converted into responses
Rather than registering a callback with the exception handler, the
application itself now has an emitter registered, and the exception
handler uses that to emit the response. It also now treats HttpException
more specially, in that they are considered to be responses to problems
that were handled, and so are not logged as errors nor passed on to the
previous exception handler.